Address RDMA public header review issues.
Keep the installed rdma-ep API self-contained by moving public provider and
queue-memory definitions into rdma-ep.h and excluding the internal
rdma-common.h helper from installation. Derive provider diagnostics from the
canonical provider names to avoid validation string drift.

Add CPU-only coverage for shared RDMA helper logic, including key
normalization and RD atomic depth selection, so regressions are caught before
hardware bring-up.

Validation:
- cmake -S . -B build
- cmake --build build --target all -j $(nproc)
- temporary install header smoke test including <rocm-xio/endpoints/rdma-ep/rdma-ep.h>
- ctest --test-dir build -R '^test-rdma-common$' --output-on-failure
- ctest --test-dir build -L unit -L rdma --output-on-failure
